flask-dashboard/requirements.txt
Flask==2.1.0
Werkzeug==2.1.1
psycopg2-binary==2.9.3
redis==4.1.0
kafka-python==2.0.2